feat: add refetching to data and refactor chroesboard

+++ src/composables/useDeleteChores.ts
@@ -1,21 +1,18 @@
+type DeleteResponse = { isDelete: boolean | null, error: string | null } | { error: Error }
+
+
 const useDeleteChores = () => {
 
-
-
-    const deleteChore = async (id: number) => {
+    const deleteChore = async (id: number): Promise<DeleteResponse> => {
         const user = useSupabaseUser();
-        if (!id || !user) return
-        try {
-            const { error } = await useFetch('/api/chores', {
-                method: 'DELETE',
-                body: { id }
-            });
-            return {
-                error
-            }
-        }
-        catch (error) {
-            console.error(error);
+        if (!id || !user) return { error: new Error('No id or user') }
+        const { error } = await useFetch('/api/chores', {
+            method: 'DELETE',
+            body: { id }
+        });
+        return {
+            isDelete: !error ? true : false,
+            error: error.value?.message || null
         }
     }
     return {
@@ -23,6 +20,5 @@ const useDeleteChores = () => {
     }
 
 }
-
 
 export default useDeleteChores
